refactor: modularize and localize SEO settings
